requirements.txt
Django
django-cors-headers
djangorestframework
whitenoise
requests
mysqlclient
mysql-connector-python